cofebem/mesh/mesh.py
import numpy as np
import meshio
import collada
from typing import Optional, Dict, List, Tuple, Union, Any, Set


class Mesh:
    """
    Mesh class for handling finite element meshes in FEM simulations.

    Attributes:
        nodes (Optional[np.ndarray]): Array of node coordinates (N x D), where N is the
            number of nodes and D is the spatial dimension.
        elements (Dict[str, np.ndarray]): Dictionary of element connectivity arrays keyed by element type.
            Example: {'triangle': np.ndarray of shape (num_elements, nodes_per_element)}
        cell_data (Dict[str, Dict[str, np.ndarray]]): Dictionary containing data arrays associated with elements,
            such as material IDs or physical tags.
        point_data (Dict[str, np.ndarray]): Dictionary containing data arrays associated with nodes.
        field_data (Dict[str, Any]): Dictionary mapping physical group names to IDs and dimensions.
        dimension (Optional[int]): Spatial dimension of the mesh (1, 2, or 3).
        element_types (List[str]): List of element types present in the mesh.
        boundary_elements (Dict[str, np.ndarray]): Dictionary of boundary elements keyed by element type.
        edges (List[Tuple[int, int]]): List of edges represented as tuples of node indices.
        faces (List[Tuple[int, ...]]): List of faces (for 3D meshes) represented as tuples of node indices.
        element_neighbors (Dict[Tuple[str, int], List[Tuple[str, int]]]): Dictionary mapping element indices to lists of neighboring elements.
        element_areas (Optional[np.ndarray]): Array of element areas (for 2D meshes).
        element_volumes (Optional[np.ndarray]): Array of element volumes (for 3D meshes).
        quality_metrics (Dict[str, np.ndarray]): Dictionary of quality metrics per element.
        partition_ids (Optional[np.ndarray]): Array assigning partition IDs to elements or nodes.
        material_ids (Optional[np.ndarray]): Array mapping elements to material IDs.
        node_dofs (Dict[int, int]): Dictionary mapping node indices to DOF indices.
        reference_element (Any): Contains shape functions and quadrature rules.
        jacobians (Optional[np.ndarray]): Array of Jacobian determinants for elements.
        generation_params (Dict[str, Any]): Parameters used to generate the mesh.
        metadata (Dict[str, Any]): Additional information about the mesh.
    """

    def __init__(self, filename: Optional[str] = None) -> None:
        """
        Initialize the Mesh object.

        Parameters:
            filename (Optional[str]): Path to the mesh file to read.
        """
        self.nodes: Optional[np.ndarray] = None
        self.elements: Dict[str, np.ndarray] = {}
        self.cell_data: Dict[str, Dict[str, np.ndarray]] = {}
        self.point_data: Dict[str, np.ndarray] = {}
        self.field_data: Dict[str, Any] = {}
        self.dimension: Optional[int] = None
        self.element_types: List[str] = []
        self.boundary_elements: Dict[str, np.ndarray] = {}
        self.edges: List[Tuple[int, int]] = []
        self.faces: List[Tuple[int, ...]] = []
        self.element_neighbors: Dict[Tuple[str, int], List[Tuple[str, int]]] = {}
        self.element_areas: Optional[np.ndarray] = None
        self.element_volumes: Optional[np.ndarray] = None
        self.quality_metrics: Dict[str, np.ndarray] = {}
        self.partition_ids: Optional[np.ndarray] = None
        self.material_ids: Optional[np.ndarray] = None
        self.node_dofs: Dict[int, int] = {}
        self.reference_element: Any = None
        self.jacobians: Optional[np.ndarray] = None
        self.generation_params: Dict[str, Any] = {}
        self.metadata: Dict[str, Any] = {}

        if filename:
            self.read(filename)

    def read(self, filename: str) -> None:
        """
        Read mesh from a file using the appropriate library.

        Parameters:
            filename (str): Path to the mesh file.
        """
        if filename.endswith(".dae"):
            self.load_collada_mesh(filename)
        else:
            self.load_meshio_mesh(filename)

    def load_collada_mesh(self, filename: str) -> None:
        """
        Read mesh from a COLLADA (.dae) file using pycollada.

        Parameters:
            filename (str): Path to the COLLADA file.
        """
        collada_mesh = collada.Collada(filename)

        # Parse nodes
        vertices = collada_mesh.geometries[0].primitives[0].vertex
        self.nodes = np.array(vertices, dtype=np.float64)

        # Parse elements
        self.elements = {}
        for geometry in collada_mesh.geometries:
            for primitive in geometry.primitives:
                if isinstance(primitive, collada.triangleset.TriangleSet):
                    indices = primitive.vertex_index
                    self.elements["triangle"] = np.array(indices, dtype=np.int64)

        # Update attributes
        self.dimension = 3 if self.nodes.shape[1] == 3 else 2
        self.element_types = list(self.elements.keys())
        self.metadata["pycollada"] = collada_mesh

    def load_meshio_mesh(self, filename: str) -> None:
        """
        Load mesh data from a meshio Mesh object.

        Parameters:
            mesh (meshio.Mesh): A meshio Mesh object.
        """
        mesh: meshio.Mesh = meshio.read(filename)
        self.nodes = mesh.points
        self.elements = {cell.type: cell.data for cell in mesh.cells}
        self.cell_data = mesh.cell_data_dict
        self.point_data = mesh.point_data
        self.field_data = mesh.field_data
        self.dimension = self.nodes.shape[1]
        self.element_types = list(self.elements.keys())
        self.metadata["meshio"] = mesh

    def write_mesh(self, filename: str, file_format: Optional[str] = None) -> None:
        """
        Write the mesh to a file using meshio.

        Parameters:
            filename (str): Path to the output mesh file.
            file_format (Optional[str]): Mesh file format. If None, inferred from filename.
        """
        # Prepare cells for meshio
        cells = [
            meshio.CellBlock(cell_type, data)
            for cell_type, data in self.elements.items()
        ]
        # Create meshio Mesh object
        mesh = meshio.Mesh(
            points=self.nodes,
            cells=cells,
            point_data=self.point_data,
            cell_data=self.cell_data,
            field_data=self.field_data,
        )
        # Write mesh to file
        meshio.write(filename, mesh, file_format=file_format)

    def get_boundary_entities(self) -> None:
        """
        Identify boundary elements or nodes.

        Populates the 'boundary_elements' attribute.
        """
        self.boundary_elements = {}
        # For simplicity, assume lower-dimensional elements represent boundaries
        lower_dim: int = self.dimension - 1 if self.dimension else 0
        boundary_element_types: List[str] = {
            1: ["vertex"],
            2: ["line", "line3"],
            3: ["triangle", "quad", "triangle6", "quad8"],
        }.get(lower_dim, [])

        for element_type in boundary_element_types:
            if element_type in self.elements:
                self.boundary_elements[element_type] = self.elements[element_type]

    def build_topology(self) -> None:
        """
        Construct relationships between nodes, edges, faces, and elements.

        Populates 'edges', 'faces', and 'element_neighbors'.
        """
        # Build edge connectivity
        self.build_edge_connectivity()

        # Build element neighbors
        self.compute_element_neighbors()

    def build_edge_connectivity(self) -> None:
        """
        Build edge connectivity from element connectivity.

        Populates the 'edges' attribute.
        """
        edge_set: Set[Tuple[int, int]] = set()
        for element_type, connectivity in self.elements.items():
            for element in connectivity:
                num_nodes: int = len(element)
                for i in range(num_nodes):
                    edge: Tuple[int, int] = tuple(
                        sorted((element[i], element[(i + 1) % num_nodes]))
                    )
                    edge_set.add(edge)
        self.edges = list(edge_set)

    def compute_element_neighbors(self) -> None:
        """
        Compute neighboring elements for each element.

        Populates the 'element_neighbors' attribute.
        """
        from collections import defaultdict

        # Initialize dictionaries
        node_to_elements: Dict[int, Set[Tuple[str, int]]] = defaultdict(set)
        self.element_neighbors = defaultdict(set)

        # Map nodes to elements
        for element_type, connectivity in self.elements.items():
            for idx, element in enumerate(connectivity):
                element_id: Tuple[str, int] = (element_type, idx)
                for node in element:
                    node_to_elements[node].add(element_id)

        # Find neighbors by shared nodes
        for element_ids in node_to_elements.values():
            for elem_a in element_ids:
                for elem_b in element_ids:
                    if elem_a != elem_b:
                        self.element_neighbors[elem_a].add(elem_b)

        # Convert sets to lists
        self.element_neighbors = {k: list(v) for k, v in self.element_neighbors.items()}

    def compute_element_areas(self) -> None:
        """
        Calculate areas of elements (for 2D meshes).

        Populates the 'element_areas' attribute.
        """
        if self.dimension != 2:
            raise ValueError("Element areas can only be computed for 2D meshes.")

        areas: List[float] = []
        for element_type, connectivity in self.elements.items():
            if element_type in ["triangle", "triangle6"]:
                for element in connectivity:
                    coords: np.ndarray = self.nodes[element]
                    area: float = 0.5 * np.linalg.norm(
                        np.cross(coords[1] - coords[0], coords[2] - coords[0])
                    )
                    areas.append(area)
            elif element_type in ["quad", "quad8"]:
                # Approximate as two triangles
                for element in connectivity:
                    coords: np.ndarray = self.nodes[element]
                    area1: float = 0.5 * np.linalg.norm(
                        np.cross(coords[1] - coords[0], coords[2] - coords[0])
                    )
                    area2: float = 0.5 * np.linalg.norm(
                        np.cross(coords[3] - coords[0], coords[2] - coords[0])
                    )
                    areas.append(area1 + area2)
            else:
                raise NotImplementedError(
                    f"Area computation not implemented for element type '{element_type}'."
                )
        self.element_areas = np.array(areas)

    def compute_element_volumes(self) -> None:
        """
        Calculate volumes of elements (for 3D meshes).

        Populates the 'element_volumes' attribute.
        """
        if self.dimension != 3:
            raise ValueError("Element volumes can only be computed for 3D meshes.")

        volumes: List[float] = []
        for element_type, connectivity in self.elements.items():
            if element_type in ["tetra", "tetra10"]:
                for element in connectivity:
                    coords: np.ndarray = self.nodes[element]
                    v: float = np.abs(
                        np.linalg.det(np.vstack((coords[1:] - coords[0], np.ones(3))))
                        / 6.0
                    )
                    volumes.append(v)
            elif element_type in ["hexahedron"]:
                # Approximate as sum of tetrahedra
                raise NotImplementedError(
                    f"Volume computation not implemented for element type '{element_type}'."
                )
            else:
                raise NotImplementedError(
                    f"Volume computation not implemented for element type '{element_type}'."
                )
        self.element_volumes = np.array(volumes)

    def visualize(self) -> None:
        """
        Visualize the mesh using PyVista.

        Note:
            Requires PyVista to be installed:
            pip install pyvista
        """
        try:
            import pyvista as pv
            from pyvista import CellType
        except ImportError:
            raise ImportError(
                "PyVista is required for visualization. Install it with 'pip install pyvista'."
            )

        # Mapping from meshio element types to PyVista cell types
        meshio_to_pyvista: Dict[str, int] = {
            "vertex": CellType.VERTEX,
            "line": CellType.LINE,
            "line3": CellType.LINE,  # Visualization of higher-order not supported
            "triangle": CellType.TRIANGLE,
            "triangle6": CellType.TRIANGLE,
            "quad": CellType.QUAD,
            "quad8": CellType.QUAD,
            "tetra": CellType.TETRA,
            "tetra10": CellType.TETRA,
            "hexahedron": CellType.HEXAHEDRON,
            "wedge": CellType.WEDGE,
            "pyramid": CellType.PYRAMID,
        }

        # Prepare cell data
        offset: int = 0
        cell_offsets: List[int] = []
        cell_types: List[int] = []
        cell_connectivity: List[int] = []

        for element_type, connectivity in self.elements.items():
            if element_type not in meshio_to_pyvista:
                print(f"Element type '{element_type}' not supported for visualization.")
                continue
            cell_type: int = meshio_to_pyvista[element_type]
            num_cells: int = connectivity.shape[0]
            num_nodes_per_cell: int = connectivity.shape[1]

            for cell in connectivity:
                # Append offset
                cell_offsets.append(offset)
                # Append connectivity (number of nodes followed by node indices)
                cell_connectivity.extend([num_nodes_per_cell])
                cell_connectivity.extend(cell.tolist())
                # Append cell type
                cell_types.append(cell_type)
                # Update offset
                offset += num_nodes_per_cell + 1

        # Create UnstructuredGrid
        cell_connectivity_array = np.array(cell_connectivity, dtype=np.int64)
        cell_offsets_array = np.array(cell_offsets, dtype=np.int64)
        cell_types_array = np.array(cell_types, dtype=np.uint8)

        grid = pv.UnstructuredGrid(
            cell_offsets_array, cell_connectivity_array, cell_types_array, self.nodes
        )
        grid.plot(show_edges=True)

    def identify_subdomains(self) -> None:
        """
        Group elements into subdomains based on physical tags.

        Populates the 'material_ids' attribute.
        """
        material_ids: List[int] = []
        cell_data = self.cell_data.get("gmsh:physical", {})
        for element_type, data in cell_data.items():
            material_ids.extend(data)
        self.material_ids = np.array(material_ids)

    def assign_materials(self, material_mapping: Dict[int, Any]) -> None:
        """
        Assign materials to subdomains or elements.

        Parameters:
            material_mapping (Dict[int, Any]): Mapping from material IDs to material properties.
        """
        self.materials: Dict[int, Any] = material_mapping

    def compute_jacobians(self) -> None:
        """
        Calculate Jacobian determinants for elements.

        Populates the 'jacobians' attribute.
        """
        jacobians: List[float] = []
        for element_type, connectivity in self.elements.items():
            for element in connectivity:
                coords: np.ndarray = self.nodes[element]
                # Compute Jacobian determinant for the element
                # Placeholder implementation
                if self.dimension == 2:
                    J = np.linalg.det(
                        np.array([coords[1] - coords[0], coords[2] - coords[0]]).T
                    )
                elif self.dimension == 3:
                    J = np.linalg.det(
                        np.array(
                            [
                                coords[1] - coords[0],
                                coords[2] - coords[0],
                                coords[3] - coords[0],
                            ]
                        ).T
                    )
                else:
                    J = np.linalg.norm(coords[1] - coords[0])
                jacobians.append(J)
        self.jacobians = np.array(jacobians)

    def evaluate_shape_functions(
        self, element_index: int, local_coords: np.ndarray
    ) -> np.ndarray:
        """
        Evaluate shape functions at given local coordinates.

        Parameters:
            element_index (int): Index of the element.
            local_coords (np.ndarray): Local coordinates where to evaluate.

        Returns:
            N (np.ndarray): Values of shape functions at local_coords.
        """
        # Placeholder implementation for linear quadrilateral element
        N = np.array([0.25, 0.25, 0.25, 0.25])
        return N

    def refine(self, elements_to_refine: List[int]) -> None:
        """
        Refine specified elements in the mesh.

        Parameters:
            elements_to_refine (List[int]): List of element indices to refine.
        """
        # Implement mesh refinement algorithm
        pass

    def coarsen(self, elements_to_coarsen: List[int]) -> None:
        """
        Coarsen specified elements in the mesh.

        Parameters:
            elements_to_coarsen (List[int]): List of element indices to coarsen.
        """
        # Implement mesh coarsening algorithm
        pass

    def partition_mesh(self, num_partitions: int) -> None:
        """
        Partition the mesh for parallel computation.

        Parameters:
            num_partitions (int): Number of partitions.
        """
        # Use a partitioning library like METIS or implement your own
        pass

    def scale(self, factor: Union[float, List[float], np.ndarray]) -> None:
        """
        Scale the mesh coordinates by a given factor.

        Parameters:
            factor (Union[float, List[float], np.ndarray]): Scaling factor(s) for each dimension.
        """
        self.nodes *= factor

    def rotate(self, angle: float, axis: str) -> None:
        """
        Rotate the mesh around a specified axis.

        Parameters:
            angle (float): Rotation angle in degrees.
            axis (str): Axis to rotate around ('x', 'y', or 'z').
        """
        from scipy.spatial.transform import Rotation as R

        rotation_axis: Dict[str, List[int]] = {
            "x": [1, 0, 0],
            "y": [0, 1, 0],
            "z": [0, 0, 1],
        }
        if axis.lower() not in rotation_axis:
            raise ValueError("Axis must be 'x', 'y', or 'z'.")
        axis_vector = rotation_axis[axis.lower()]
        r = R.from_rotvec(np.radians(angle) * np.array(axis_vector))
        self.nodes = r.apply(self.nodes)

    def translate(self, vector: Union[List[float], np.ndarray]) -> None:
        """
        Translate the mesh by a given vector.

        Parameters:
            vector (Union[List[float], np.ndarray]): Translation vector.
        """
        self.nodes += vector

    def compute_quality_metrics(self) -> None:
        """
        Calculate quality metrics for each element.

        Populates the 'quality_metrics' attribute.
        """
        # Placeholder implementation
        num_elements: int = sum(len(conn) for conn in self.elements.values())
        self.quality_metrics["aspect_ratio"] = np.ones(num_elements)

    def export_mesh(self, filename: str, format: str) -> None:
        """
        Export the mesh in various formats.

        Parameters:
            filename (str): Path to the output file.
            format (str): Format to export (e.g., 'xdmf', 'vtk').
        """
        self.write_mesh(filename, file_format=format)

    def build_node_adjacency(self) -> None:
        """
        Construct adjacency lists for nodes.

        Populates the 'node_neighbors' attribute.
        """
        from collections import defaultdict

        node_neighbors: Dict[int, Set[int]] = defaultdict(set)
        for element_type, connectivity in self.elements.items():
            for element in connectivity:
                for i, node in enumerate(element):
                    for neighbor in element:
                        if node != neighbor:
                            node_neighbors[node].add(neighbor)
        self.node_neighbors: Dict[int, List[int]] = {
            k: list(v) for k, v in node_neighbors.items()
        }

    def load_from_geometry(self, geometry_file: str) -> None:
        """
        Create a mesh based on a geometry file.

        Parameters:
            geometry_file (str): Path to the geometry file.
        """
        # Implement mesh generation from geometry
        pass

    # Additional methods as needed...
